moved some code away from msc_drive
